Tidy bulk signature verifier
